feat: Add method to send password recovery email
Add a new method in the EmailService class to send a password recovery email to a user. The method takes a UserEntity object and a token as parameters and constructs an email with the recovery link. The email is then sent using the configured mail sender.

Refactor the EmailService class to include the new method for sending password recovery emails.